feat: reorganizing, refactor platform interface
