+++ packages/audioplayers_platform_interface/lib/map_extension.dart
@@ -1,0 +1,15 @@
+extension MapParser on Map<dynamic, dynamic> {
+  bool containsKey(String key) => this.containsKey(key);
+
+  String getString(String key) {
+    return this[key] as String;
+  }
+
+  int getInt(String key) {
+    return this[key] as int;
+  }
+
+  bool getBool(String key) {
+    return this[key] as bool;
+  }
+}
