+++ packages/audioplayers_platform_interface/lib/method_channel_extension.dart
@@ -9,19 +9,3 @@ extension StandardMethodChannel on MethodChannel {
     return invokeMethod<T>(method, args);
   }
 }
-
-extension MapParser on Map<dynamic, dynamic> {
-  bool containsKey(String key) => this.containsKey(key);
-
-  String getString(String key) {
-    return this[key] as String;
-  }
-
-  int getInt(String key) {
-    return this[key] as int;
-  }
-
-  bool getBool(String key) {
-    return this[key] as bool;
-  }
-}
